Some work

+++ sensor/sensor.h
@@ -22,12 +22,8 @@ namespace Sensor {
 
   class Vfs {
   public:
-//    uint8_t motion_occurred;
-//    uint8_t motion_overflow;
-    Cartesian2<int8_t> motion;
+    Cartesian2<float> motion;
     uint8_t surface_quality;
-//    uint16_t shutter;
-//    uint8_t max_pixel_value;
 
     friend std::ostream& operator<<(std::ostream& os, const Vfs& vfs);
   };
